Separate axis into multiple components
