+++ src/lib/plot/axis/axis-line.js
@@ -1,0 +1,69 @@
+// Copyright (c) 2016 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+import React from 'react';
+import {ORIENTATION} from '../../utils/axis-utils';
+
+const {LEFT, RIGHT, TOP, BOTTOM} = ORIENTATION;
+
+const propTypes = {
+  width: React.PropTypes.number.isRequired,
+  height: React.PropTypes.number.isRequired,
+  orientation: React.PropTypes.oneOf([
+    LEFT, RIGHT, TOP, BOTTOM
+  ]).isRequired
+};
+
+function AxisLine({orientation, width, height}) {
+  let x1;
+  let x2;
+  let y1;
+  let y2;
+  if (orientation === LEFT) {
+    x1 = width;
+    x2 = width;
+    y1 = 0;
+    y2 = height;
+  } else if (orientation === RIGHT) {
+    x1 = 0;
+    x2 = 0;
+    y1 = 0;
+    y2 = height
+  } else if (orientation === TOP) {
+    x1 = 0;
+    x2 = width;
+    y1 = height;
+    y2 = height;
+  } else {
+    x1 = 0;
+    x2 = width;
+    y1 = 0;
+    y2 = 0;
+  }
+  return (
+    <line x1={x1} x2={x2} y1={y1} y2={y2} className="rv-xy-plot__axis__line"/>
+  );
+}
+
+AxisLine.displayName = 'AxisLine';
+AxisLine.propTypes = propTypes;
+
+export default AxisLine;
+
