+++ src/test/components.js
@@ -29,8 +29,8 @@ import MarkSeries from '../lib/plot/series/mark-series';
 import HeatmapSeries from '../lib/plot/series/heatmap-series';
 import VerticalBarSeries from '../lib/plot/series/vertical-bar-series';
 import HorizontalBarSeries from '../lib/plot/series/horizontal-bar-series';
-import XAxisBottom from '../lib/plot/x-axis';
-import YAxisLeft from '../lib/plot/y-axis';
+import XAxisBottom from '../lib/plot/axis/x-axis';
+import YAxisLeft from '../lib/plot/axis/y-axis';
 import VerticalGrid from '../lib/plot/vertical-grid-lines';
 import HorizontalGrid from '../lib/plot/horizontal-grid-lines';
 import XYPlot from '../lib/plot/xy-plot';
