+++ src/test/xy-plot.js
@@ -22,7 +22,7 @@ import test from 'tape';
 import React from 'react';
 import {shallow} from 'enzyme';
 import VerticalBarSeries from '../lib/plot/series/vertical-bar-series';
-import XAxis from '../lib/plot/x-axis';
+import XAxis from '../lib/plot/axis/x-axis';
 import XYPlot from '../lib/plot/xy-plot';
 
 test('Render a stacked bar chart', assert => {
